Made the code that gets user names more selective.
This update makes it so that the user name is in most cases obtained
from the link (`a`) rather than from the link's parent. For one thing,
this removes the need to cleanup after obtaining the text. That cleanup
was probably buggy anyway. (User names can contain numbers.) Also, the
method previously used interfered with other user scripts that add
information next to the user name. The updated code is more conservative
and thus will play better with other user scripts.